# Significant Optimizations: fasthttp Integration and Memory Allocation Reduction