--- a/.gitignore
+++ b/.gitignore
@@ -23,3 +23,7 @@ go.work.sum
 
 # env file
 .env
+
+# IDE files
+.idea/
+.vscode/