## Summary
This PR optimises the library by:
- Replacing net/http with fasthttp for faster HTTP requests and lower memory usage.
- Introducing jsoniter for faster JSON unmarshalling.
- Using buffer pooling to reduce memory allocations.

## Benchmark Results
Before update:
- QueryIP: 306,275,823 ns/op, 54,724 B/op, 93 allocs/op
- QueryOwnIP: 27,697 ns/op, 5,578 B/op, 66 allocs/op

After update:
- QueryIP: 296,105,344 ns/op, 4,232 B/op, 27 allocs/op
- QueryOwnIP: 19,984 ns/op, 1,603 B/op, 18 allocs/op